+++ modules/base/index.ts
@@ -1,10 +1,6 @@
 export { default as extendStyles } from "./extendStyles";
 export { default as useStyles } from "./useStyles";
 export { default as useTheme } from "./useTheme";
-export {
-  default as withStyles,
-  withStylesHOC,
-  WithStylesType
-} from "./withStyles";
+export { default as withStyles, withStylesHOC } from "./withStyles";
 export { default as ComponentRegistry } from "./Registry";
 export * from "./setup";
