+++ modules/tokens/index.scss
@@ -1,6 +1,5 @@
 :root {
   --color-primary: #00321e;
-  --color-secondary: ;
   --color-alert-alert25: #d86464;
   --color-alert-alert50: #d86464;
   --color-alert-alert100: #d86464;
@@ -13,18 +12,16 @@
   --color-white: #ffffff;
   --color-black: #070707;
   --color-background-background25: #f8f8f8;
-  --color-background-background50: ;
   --color-background-background100: #161616;
   --color-background-overlay: rgba(4, 4, 15, 0.7);
   --color-grey-grey25: #f6f6f6;
   --color-grey-grey50: #cccccc;
   --color-grey-grey100: #808080;
-  --color-grey-greenish: E5EBE8;
 
   --space-unit: 16px;
   --space-xxs: 4px;
   --space-xs: 8px;
-  --space-sm: 10.666666666666666px;
+  --space-sm: 12px;
   --space-md: 16px;
   --space-lg: 24px;
   --space-xl: 32px;
